requirements.txt
streamlit==1.22.0
numpy==1.24.3
Pillow==9.5.0
cryptography==41.0.1
matplotlib==3.7.1